requirements.txt
absl-py==1.4.0
addict==2.4.0
altair==4.2.2
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
appnope==0.1.3
array-record==0.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.0.12
clearml==1.10.4
clearml-serving==1.3.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
dm-tree==0.1.8
easyocr==1.6.2
editdistance==0.6.2
efficientnet==1.0.0
entrypoints==0.4
essential-generators==1.0
etils==1.2.0
executing==1.2.0
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
furl==2.1.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
ipython==8.13.2
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.9.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lmdb==1.4.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mmcv==2.0.0
mmdet==3.0.0
mmengine==0.7.3
model-index==0.1.11
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
openmim==0.3.7
opt-einsum==3.3.0
ordered-set==4.1.0
orderedmultidict==1.0.1
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.5.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycocotools==2.0.6
pydantic==1.9.2
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.4.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pystac==1.4.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
radiant-mlhub==0.5.5
rapidfuzz==3.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
Shapely==1.8.5.post1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.0
stack-data==0.6.2
streamlit==1.22.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metadata==1.13.1
tensorflow-metal==0.8.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.1
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
validators==0.20.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.3.3
wrapt==1.14.1
yapf==0.33.0
zipp==3.15.0
